auction_server/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
base58==2.1.1
boto3==1.21.46
botocore==1.24.46
click==8.1.2
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.4
django-allauth==0.50.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
filelock==3.6.0
Flask==2.1.1
Flask-Cors==3.0.10
greenlet==1.1.2
importlib-metadata==4.11.3
ipfshttpclient==0.7.0
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
multiaddr==0.0.9
netaddr==0.8.0
packaging==21.3
Pillow==9.1.0
platformdirs==2.5.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyparsing==3.0.8
pytesseract==0.3.9
python3-openid==3.2.0
requests-oauthlib==1.3.1
s3transfer==0.5.2
sqlparse==0.4.2
tk==0.1.0
toml==0.10.2
varint==1.0.2
virtualenv==20.14.0
Werkzeug==2.1.1
zipp==3.8.0
